+++ etc/s6-overlay/s6-rc.d/cron/run
@@ -6,9 +6,9 @@ DEFAULT_CRON="0 20 * * *"
 sed -i '$d' /etc/crontabs/root
 
 if [[ -z ${CRON} ]]; then
-    echo "${DEFAULT_CRON} /usr/local/bin/python3 -u /app/main.py" >> /etc/crontabs/root
+    echo "${DEFAULT_CRON} cd /app && python3 -m tmdb_notifier" >> /etc/crontabs/root
 else
-    echo "${CRON} /usr/local/bin/python3 -u /app/main.py" >> /etc/crontabs/root
+    echo "${CRON} cd /app && python3 -m tmdb_notifier" >> /etc/crontabs/root
 fi
 
 exec /usr/sbin/crond -f -S -l 5
