+++ etc/s6-overlay/s6-rc.d/run-app/run
@@ -1,4 +1,4 @@
 #!/usr/bin/with-contenv bash
 # shellcheck shell=bash
 
-/usr/local/bin/python3 -u /app/main.py
+cd /app && python3 -m tmdb_notifier
