--- src/models/__init__.py
@@ -1,3 +1,0 @@
-from src.models.movie import Movie
-
-__all__ = ["Movie"]
