--- src/notifiers/__init__.py
@@ -1,4 +1,0 @@
-from src.notifiers.base import Notifier
-from src.notifiers.notifiers import Notifiers
-
-__all__ = ["Notifier", "Notifiers"]
