+++ tmdb_notifier/Session.py
