+++ tmdb_notifier/models/__init__.py
@@ -1,0 +1,3 @@
+from tmdb_notifier.models.movie import Movie
+
+__all__ = ["Movie"]
