+++ tmdb_notifier/notifiers/__init__.py
@@ -1,0 +1,4 @@
+from tmdb_notifier.notifiers.base import Notifier
+from tmdb_notifier.notifiers.notifiers import Notifiers
+
+__all__ = ["Notifier", "Notifiers"]
